Engine refactor
Due to the need to handle a higher number of collision objects, a new form of collision management is being put forward.  Previously, each object was responsible for checking itself against every other object to detect collisions.  Thereby running many more checks than actually is needed.  The new code seeks to handle all collisions with a collision manager.  An object updates the collision manager if it moves, and the collision manger runs but once per frame to determine collisions.  However, testing is needed to see what shortcomings and improvements this new system brings.  ps.  You will probably need to edit the path in order for this to run on your computer.